requirements.txt
asgiref==3.4.1                                                                                                                                                                                                                                                                                                             
certifi==2022.12.7                                                                                                                                                                                                                                                                                                         
charset-normalizer==2.0.12                                                                                                                                                                                                                                                                                                 
Django==3.2.18                                                                                                                                                                                                                                                                                                             
django-environ==0.10.0                                                                                                                                                                                                                                                                                                     
django-ranged-response==0.2.0                                                                                                                                                                                                                                                                                              
django-simple-captcha==0.5.17   
django-bootstrap5==23.3                                                                                                                                                                                                                                                                                           
idna==3.4                                                                                                                                                                                                                                                                                                                  
Pillow==8.4.0                                                                                                                                                                                                                                                                                                              
pytz==2023.3                                                                                                                                                                                                                                                                                                               
requests==2.27.1                                                                                                                                                                                                                                                                                                           
sqlparse==0.4.3                                                                                                                                                                                                                                                                                                            
typing_extensions==4.1.1                                                                                                                                                                                                                                                                                                   
urllib3==1.26.15                                                                                                                                                                                                                                                                                                           
vk-api==11.9.9                                                                                                                                                                                                                                                                                                             
whitenoise==5.3.0
djangorestframework==3.12.4
drf-extra-fields==3.4.1
python-dotenv==0.19.0

gunicorn==20.1.0

pytest==6.2.4
pytest-pythonpath==0.7.3
pytest-django==4.4.0
pytest-lazy-fixture==0.6.3

#drf-yasg==1.21.7
#flake8==6.1.0
#isort==5.12.0
#autopep8==2.0.4
#django_debug_toolbar==3.8.1
